dkoltovich/lesson_29/requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
celery==5.2.7
Django==4.1.7
django-celery-beat==2.5.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2